# eidStack-CMU/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /short-url/create:
    post:
      operationId: ShortUrlController_createShortUrl
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShortUrlDto'
      responses:
        '201':
          description: ''
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Create short URL
      tags: &ref_1
        - short-url
  /short-url/resolve:
    get:
      operationId: ShortUrlController_getOriginalUrl
      parameters:
        - name: code
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Resolve code to original URL
      tags: *ref_1
  /short-url/s/{code}:
    get:
      operationId: ShortUrlController_redirectToOriginal
      parameters:
        - name: code
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags: *ref_1
...
  /verification/proofStatus:
    get:
      operationId: VerificationController_getProofStatus
      parameters:
        - name: proofRecordId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Proof state (string)
        '400':
          description: Invalid proofRecordId
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Get proof/request status
      tags: &ref_3
        - verification
  /verification/createProofRequest:
    post:
      operationId: VerificationController_createProofRequest
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProofRequestDto'
      responses:
        '201':
          description: Proof request created (object)
        '400':
          description: Validation error
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Create proof request (OOB or connection-based)
      tags: *ref_3
info:
  title: e-ID REST API
  description: REST APIs converted from GraphQL
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateShortUrlDto:
      type: object
      properties:
        originalUrl:
          type: string
      required:
        - originalUrl
...
    RestrictionDto:
      type: object
      properties:
        schema:
          type: object
          description: Schema filter / predicate
        credDefId:
          type: object
          description: Credential definition id filter
    AttributeRequestDto:
      type: object
      properties:
        name:
          type: string
          description: Attribute name to request (e.g. email, given_name)
        restrictions:
          description: Optional array of restrictions (schema, credDefId)
          type: array
          items:
            $ref: '#/components/schemas/RestrictionDto'
      required:
        - name
    PredicateRequestDto:
      type: object
      properties:
        name:
          type: string
          description: Predicate attribute name
        pType:
          type: string
          description: pType (e.g. '>=', '<=', '>')
        pValue:
          type: number
          description: predicate value
      required:
        - name
        - pType
        - pValue
    ProofRequestDto:
      type: object
      properties:
        credDefId:
          type: string
          description: Credential definition id used for this proof request
        attributes:
          description: Attributes to request
          type: array
          items:
            $ref: '#/components/schemas/AttributeRequestDto'
        predicates:
          description: Optional predicates
          type: array
          items:
            $ref: '#/components/schemas/PredicateRequestDto'
        comment:
          type: string
          description: Optional comment to include in proof request
        connectionId:
          type: string
          description: Optional connectionId if request should be connection-based
      required:
        - credDefId
        - attributes
    ErrorResponse:
      type: object
      properties:
        correlationId:
          type: string
        statusCode:
          type: integer
          format: int32
        error:
          type: string
        message:
          oneOf:
            - type: string
            - type: array
            - type: object